+++ lib/src/framework/export.dart
@@ -20,5 +20,6 @@ export 'collection/sortable_hash_set.dart';
 export 'configuration.dart';
 export 'draw/overlay.dart';
 export 'grid_tick_mode.dart';
+export 'interaction_event.dart';
 export 'logger.dart';
 export 'serializers.dart';